+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - (Feature) Immutable spec
 - (Bugfix) Proper agent cleanout
 - (Bugfix) Fix ClusterScaling integration
+- (Feature) Sensitive information protection
 
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long